+++ pinery-miso/src/it/resources/db/migration/pinery-miso_plain_test_data.sql
@@ -215,7 +215,7 @@ INSERT INTO Indices (indexId, indexFamilyId, name, sequence, position) VALUES
 -- Plain sample data
 
 INSERT INTO Project(projectId, name, alias, shortName, created, description,
-  progress, referenceGenomeId, lastModified, creator, lastModifier) VALUES
+  status, referenceGenomeId, lastModified, creator, lastModifier) VALUES
   (1, 'PRO1', 'Project One', 'PRO1', '2017-06-27', 'integration test project one', 'ACTIVE', 1, '2017-06-27 14:11:00', 1, 1),
   (2, 'PRO2', 'Project Two', 'PRO2', '2017-06-27', 'integration test project two', 'ACTIVE', 1, '2017-06-27 14:11:00', 1, 1);